reorg bsync server routes
