+++ packages/bsync/src/routes/scan-mute-operations.ts
@@ -1,0 +1,69 @@
+import { once } from 'node:events'
+import { Code, ConnectError, ServiceImpl } from '@connectrpc/connect'
+import { Service } from '../gen/bsync_connect'
+import { ScanMuteOperationsResponse } from '../gen/bsync_pb'
+import AppContext from '../context'
+import { createMuteOpChannel } from '../db/schema/mute_op'
+import { authWithApiKey } from './auth'
+
+export default (ctx: AppContext): Partial<ServiceImpl<typeof Service>> => ({
+  async scanMuteOperations(req, handlerCtx) {
+    authWithApiKey(ctx, handlerCtx)
+    const { db, events } = ctx
+    const limit = req.limit || 1000
+    const cursor = validCursor(req.cursor)
+    const nextMuteOpPromise = once(events, createMuteOpChannel, {
+      signal: AbortSignal.timeout(10000),
+    })
+    nextMuteOpPromise.catch(() => null) // ensure timeout is always handled
+
+    const nextMuteOpPageQb = db.db
+      .selectFrom('mute_op')
+      .selectAll()
+      .where('id', '>', cursor ?? -1)
+      .orderBy('id', 'asc')
+      .limit(limit)
+
+    let ops = await nextMuteOpPageQb.execute()
+
+    if (!ops.length) {
+      // if there were no ops on the page, wait for an event then try again.
+      try {
+        await nextMuteOpPromise
+      } catch (err) {
+        return new ScanMuteOperationsResponse({
+          operations: [],
+          cursor: req.cursor,
+        })
+      }
+      ops = await nextMuteOpPageQb.execute()
+      if (!ops.length) {
+        return new ScanMuteOperationsResponse({
+          operations: [],
+          cursor: req.cursor,
+        })
+      }
+    }
+
+    const lastOp = ops[ops.length - 1]
+
+    return new ScanMuteOperationsResponse({
+      operations: ops.map((op) => ({
+        id: op.id.toString(),
+        type: op.type,
+        actorDid: op.actorDid,
+        subject: op.subject,
+      })),
+      cursor: lastOp.id.toString(),
+    })
+  },
+})
+
+const validCursor = (cursor: string): number | null => {
+  if (cursor === '') return null
+  const int = parseInt(cursor, 10)
+  if (isNaN(int) || int < 0) {
+    throw new ConnectError('invalid cursor', Code.InvalidArgument)
+  }
+  return int
+}
